# frontend/ticketing_app_ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/users/me/contact-info": {
      "put": {
        "tags": [
          "User Management"
        ],
        "summary": "Aggiorna le informazioni di contatto dell'utente corrente",
        "description": "Permette all'utente di aggiornare il proprio numero di telefono e codice fiscale.",
        "operationId": "updateContactInfo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserContactUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Utente non trovato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "200": {
            "description": "Informazioni di contatto aggiornate con successo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "400": {
            "description": "Dati di input non validi.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{userId}/role": {
      "patch": {
        "tags": [
          "User Management"
        ],
        "summary": "Aggiorna il ruolo di un utente",
        "description": "Permette agli ADMIN di modificare il ruolo di un utente esistente. Non è possibile modificare il ruolo di altri ADMIN.",
        "operationId": "updateUserRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID dell'utente da aggiornare",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "uuid-user-1"
          },
          {
            "name": "newRole",
            "in": "query",
            "description": "Il nuovo ruolo da assegnare",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ADMIN",
                "HELPER_JUNIOR",
                "HELPER_SENIOR",
                "PM",
                "USER"
              ]
            },
            "example": "HELPER_JUNIOR"
          }
        ],
        "responses": {
          "404": {
            "description": "Utente non trovato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "400": {
            "description": "Dati di input non validi (es. ruolo già presente).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "403": {
            "description": "Non autorizzato (solo ADMIN o tentativo di modificare un altro ADMIN).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "200": {
            "description": "Ruolo utente aggiornato con successo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "User Management"
        ],
        "summary": "Recupera tutti gli utenti (tranne se stessi)",
        "description": "Permette agli ADMIN di visualizzare tutti gli utenti registrati nel sistema, escluso il proprio account.",
        "operationId": "getAllUsers",
        "responses": {
          "200": {
            "description": "Lista utenti recuperata con successo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Non autorizzato (solo ADMIN).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "User Management"
        ],
        "summary": "Recupera i dettagli dell'utente corrente",
        "description": "Recupera le informazioni del profilo dell'utente autenticato.",
        "operationId": "getMe",
        "responses": {
          "200": {
            "description": "Dettagli utente recuperati con successo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "404": {
            "description": "Utente non trovato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/helpers": {
      "get": {
        "tags": [
          "User Management"
        ],
        "summary": "Recupera tutti gli Helper (Junior, Senior) e PM",
        "description": "Permette agli ADMIN di visualizzare tutti gli utenti con ruolo Helper_Junior, Helper_Senior o PM.",
        "operationId": "getHelpers",
        "responses": {
          "200": {
            "description": "Lista Helper/PM recuperata con successo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Non autorizzato (solo ADMIN).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/helpers-junior": {
      "get": {
        "tags": [
          "User Management"
        ],
        "summary": "Recupera tutti gli Helper Junior",
        "description": "Permette agli ADMIN di visualizzare tutti gli utenti con ruolo Helper_Junior.",
        "operationId": "getHelpersJunior",
        "responses": {
          "401": {
            "description": "Non autenticato.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Lista Helper Junior recuperata con successo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Non autorizzato (solo ADMIN).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/helpers-and-admins": {
      "get": {
        "tags": [
          "User Management"
        ],
        "summary": "Recupera tutti gli Helper, PM e Admin",
        "description": "Permette a Helper, PM e Admin di visualizzare tutti gli utenti con ruolo Helper, PM o Admin.",
        "operationId": "getHelpersAndAdmins",
        "responses": {
          "401": {
            "description": "Non autenticato.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Non autorizzato.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Lista Helper/PM/Admin recuperata con successo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{userId}": {
      "delete": {
        "tags": [
          "User Management"
        ],
        "summary": "Elimina un account utente",
        "description": "Permette agli ADMIN di eliminare un account utente e tutti i ticket associati. Non è possibile eliminare altri ADMIN.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID dell'utente da eliminare",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "uuid-user-1"
          }
        ],
        "responses": {
          "404": {
            "description": "Utente non trovato."
          },
          "204": {
            "description": "Account utente eliminato con successo."
          },
          "401": {
            "description": "Non autenticato."
          },
          "403": {
            "description": "Non autorizzato (solo ADMIN o tentativo di eliminare un altro ADMIN)."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserContactUpdateRequest": {
        "type": "object",
        "properties": {
          "fiscalCode": {
            "pattern": "^[A-Z]{6}[0-9]{2}[A-Z][0-9]{2}[A-Z][0-9]{3}[A-Z]$",
            "type": "string"
          },
          "phoneNumber": {
            "pattern": "^\\+?[0-9]{7,15}$",
            "type": "string"
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "HELPER_JUNIOR",
              "HELPER_SENIOR",
              "PM",
              "USER"
            ]
          },
          "fiscalCode": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "keycloak": {
        "type": "oauth2",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "flows": {
          "password": {
            "authorizationUrl": "http://localhost:9090/realms/ticket_tool/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:9090/realms/ticket_tool/protocol/openid-connect/token",
            "scopes": {

            }
          }
        }
      }
    }
  }
}